Split FieldDefinition into three concepts
The old FieldDefinition defined everything about how a field was parsed and
how it could be used in a query but these bits of information came from
different sources so it was confusing. Now there are three things:

FieldReference - A field name and a boost. This is all you can get from
parsing a field.
FieldDefinition - Stuff about a field that must be set up outside of the
query itself. For now thats just the name of the field to use for quoted
terms.
FieldUsage - The field information required to actually build a query. You
get this by calling FieldsHelper.resolve so its safe to assume that
FieldUsages always have the aliases expanded and whitelist and blacklist
checked.