Support rewriting prefix queries
If you have a prefix term built with an edgeNGram filter you can rewrite
prefix queries against it. This happens by default if you name the field
field_name.prefix_precise. It is so named because prefix queries should run
against the "precise" analysis of the field rather than the "standard",
stemmed analysis.