Regex support
Adds support for trigram accelerated regexes if the wikimedia-extra plugin
is on the classpath. If the plugin isn't on the classpath and the query tries
to turn regexes on by setting the "regex" config section the query will fail.
If regexes aren't enabled and the user tries to use them they'll just be
interpreted as a term query containing /s.

Needs https://gerrit.wikimedia.org/r/#/c/215913/